Rename migrations to cli
