+++ saffier/cli/operations/current.py
@@ -1,7 +1,7 @@
 import click
 
-from saffier.migrations.base import current as _current
-from saffier.migrations.env import MigrationEnv
+from saffier.cli.base import current as _current
+from saffier.cli.env import MigrationEnv
 
 
 @click.option(
