+++ saffier/cli/operations/list_templates.py
@@ -3,7 +3,7 @@ Client to interact with Saffier models and migrations.
 """
 import click
 
-from saffier.migrations.base import list_templates as template_list
+from saffier.cli.base import list_templates as template_list
 
 
 @click.command(name="list-templates")
