+++ saffier/cli/operations/makemigrations.py
@@ -1,7 +1,13 @@
+"""
+Client to interact with Saffier models and migrations.
+"""
+
+from typing import Any
+
 import click
 
-from saffier.migrations.base import revision as _revision
-from saffier.migrations.env import MigrationEnv
+from saffier.cli.base import migrate as _migrate
+from saffier.cli.env import MigrationEnv
 
 
 @click.option(
@@ -11,14 +17,6 @@ from saffier.migrations.env import MigrationEnv
     help=('Migration script directory (default is "migrations")'),
 )
 @click.option("-m", "--message", default=None, help="Revision message")
-@click.option(
-    "--autogenerate",
-    is_flag=True,
-    help=(
-        "Populate revision script with candidate migration "
-        "operations, based on comparison of database to model"
-    ),
-)
 @click.option(
     "--sql", is_flag=True, help=("Don't emit SQL to database - dump to standard output " "instead")
 )
@@ -39,29 +37,24 @@ from saffier.migrations.env import MigrationEnv
 @click.option(
     "--rev-id", default=None, help=("Specify a hardcoded revision id instead of generating " "one")
 )
+@click.option(
+    "-x", "--arg", multiple=True, help="Additional arguments consumed by custom env.py scripts"
+)
 @click.command()
-def revision(
+def makemigrations(
     env: MigrationEnv,
     directory: str,
     message: str,
-    autogenerate: bool,
     sql: bool,
     head: str,
     splice: bool,
     branch_label: str,
     version_path: str,
     rev_id: str,
+    arg: Any,
 ) -> None:
-    """Create a new revision file."""
-    _revision(
-        env.app,
-        directory,
-        message,
-        autogenerate,
-        sql,
-        head,
-        splice,
-        branch_label,
-        version_path,
-        rev_id,
+    """Autogenerate a new revision file (Alias for
+    'revision --autogenerate')"""
+    _migrate(
+        env.app, directory, message, sql, head, splice, branch_label, version_path, rev_id, arg
     )
