+++ saffier/cli/operations/shell/enums.py
