+++ saffier/cli/operations/show.py
@@ -1,7 +1,7 @@
 import click
 
-from saffier.migrations.base import show as _show
-from saffier.migrations.env import MigrationEnv
+from saffier.cli.base import show as _show
+from saffier.cli.env import MigrationEnv
 
 
 @click.option(
