+++ saffier/cli/templates/default/env.py
@@ -12,9 +12,9 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.asyncio import create_async_engine
 
 from saffier import settings
+from saffier.cli.constants import APP_PARAMETER
+from saffier.cli.env import MigrationEnv
 from saffier.exceptions import SaffierException
-from saffier.migrations.constants import APP_PARAMETER
-from saffier.migrations.env import MigrationEnv
 
 # The console used for the outputs
 console = Console()
