+++ saffier/cli/templates/default/script.py.mako
